requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
PyPDF2
pandas==2.2.2
langchain-google-genai==1.0.7
langchain==0.2.7
faiss-cpu
